# Optimize HashAgg for sparse HHashTable iteration