@hlinnaka 2017-04-04
I'm not sure I understand this. The hash keys are assumed to be distributed more or less equally across all buckets. Doesn't that mean that you will very quickly have at least one bucket in each segment? With `MINIMUM_SEGSIZE` of 1024, when roughly only 0.1% of the hash table is used, you will have on average one bucket in each segment, and you no longer benefit. Or to put another way, this only helps, if the planner got the row estimates off by a factor of 1000 or more. I'm sure that happens sometimes, but seems like a pretty narrow case.

Did you consider simply starting with a smaller hash table, and enlarging it when it when it fills up? That's what you usually do with hash tables, if you don't know the optimal size ahead of time...