@hlinnaka 2017-04-04
I.e. how about doing the that was done in upstream commit 30d7ae3c76d2de144232ae6ab328ca86b70e72c3?