@hardikar 2017-04-04
> Did you consider simply starting with a smaller hash table, and enlarging it when it when it fills up? > That's what you usually do with hash tables, if you don't know the optimal size ahead of time...

I did consider growing the table to solve this issue ... This PR was an attempt to change as little as possible to solve those small set of cases. But, I agree with you, implementing a dynamic hash table is the correct approach to this. I just send out an e-mail to gpdb-dev exploring how to move forward in that case, since that will require some proposal/discussion.