@hardikar 2017-04-14
Closing this PR in favor of https://github.com/greenplum-db/gpdb/pull/2244.